// src/app/models/consultas.ts
export interface Consultas {
  caminhoneiro: string;
  reclamacao: string;
  tipoSanquineo: string;
  fumante: boolean;
  obeso: boolean;
  drogas: boolean;
  alcool: boolean;
  coracao: boolean;
  diabetico: boolean;
  cirurgia: boolean;
  insonia: boolean;
  exercicios: boolean;
  fisico: boolean;
  psicologico: boolean;
  especialidade: string;
}